ROS: Refactor node
